+++ test/widgets/inventory_list_item_test.dart
